Support double symbolic links in /etc/localtime
According to the file hierachy, /etc may be read-only, however in
this case there is no way to set the time-zone.

This patch allows the system to have a symbolic link to a second
symbolic link in a writable filesystem. The second symbolic link
will contain the timezone name.

This patch does not allow infinite recursion because of the danger
of link-loops, and because it becomes harder to guaruantee that
no link in the chain will fall under the control of a non-
privaleged user. Ultimately, endless recursion is unlikely to
solve a real use-case.