path-util: Added path_normalize
Normalize a pathname by collapsing redundant separators and
up-level references so that A//B, A/B/, A/./B and A/foo/../B all
become A/B.

In a relative path, all ../ up-level references will be collected
together at the start of the string.

This string manipulation may change the meaning of a path that
contains symbolic links.